Allow specifying a redis URL directly
This allows an admin to specify a custom redis URL overriding the dynamically built one. This is meant for a couple of more advanced use cases:

* Allows specifying a UNIX socket path to Redis
* Can use a specific user name (rather than the `default` user) allowing for more granular ACLs in Redis 6+

As a side effect, any future keyword arguments to `redis-py` can be passed in immediately before there are specific environment variables for each of them.

Signed-off-by: No GUI <14547+nogweii@users.noreply.github.com>