web: revised package.json (#5612)
* FOUND: Revise the package.json file

Reviewed the package.json file and separated out the dependencies
from the devDependencies.  Normally, this doesn't make a difference,
but when deployed as libraries the devDependencies don't normally
get considered by the deploying aplication.

* FIX Undo overrides and work-arounds for bad upstream postgres build.

* 5171: Fixed README to comply with Prettier rules.

I'm pretty impressed that that worked.  Good on Jens for having that
in the prettier rules.

* web: revised package.json

Removed the migration and web/README.md file.  The former should not have
been included; the latter is currently unprofessional in tone.

* fix docker build failing as we set NODE_ENV but need dev dependencies to build

Signed-off-by: Jens Langhammer <jens@goauthentik.io>

---------

Signed-off-by: Jens Langhammer <jens@goauthentik.io>
Co-authored-by: Jens Langhammer <jens@goauthentik.io>