web: unravel the search-select for certificates
This commit applies the same procedure and processes to handling CertficateKeyPairs as was applied
to the search for flows.  All 12 of these searches were the same, with these notable differences: the
key IDs are often stored in differently named fields, so that couldn't be routinized; some of these
searches only apply to fields where the key is defined, so the boolean "has-key" is available; one
specific search could match multiple keys in some certificates, so a special case has been added
for it.